requirements.txt
requests
biopython
pandas
python-dotenv
